hello/requirements.txt
## Top level dependencies
alembic==0.8.10
bleach==2.0.0
celery==4.0.2
Flask-Admin==1.4.2
Flask-DebugToolbar==0.10.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
Flask==0.12
humanize==0.5.1
Jinja2==2.9.5
mysqlclient==1.3.10
passlib==1.7.0
Pillow==4.0.0
python-magic==0.4.12
SQLAlchemy==1.1.5
Werkzeug==0.11.15
WTForms==2.1
##

## Top level asset dependencies
cssmin==0.2.0
Flask-Assets==0.12
jsmin==2.2.1
libsass==0.12.3
## - end

## Downstream dependencies
##


## celery dependencies
amqp==2.1.4
billiard==3.5.0.2
kombu==4.0.2
vine==1.1.3

## bleach dependencies
html5lib==0.999999999
webencodings==0.5
